+++ feature-push-notifications/src/main/java/io/novafoundation/nova/feature_push_notifications/data/data/PushNotificationsService.kt
@@ -1,8 +1,12 @@
 package io.novafoundation.nova.feature_push_notifications.data.data
 
+import android.util.Log
+import com.google.android.gms.tasks.OnCompleteListener
+import com.google.firebase.messaging.FirebaseMessaging
 import io.novafoundation.nova.common.data.storage.Preferences
 import io.novafoundation.nova.common.utils.coroutines.RootScope
 import io.novafoundation.nova.common.utils.repeatUntil
+import io.novafoundation.nova.feature_push_notifications.BuildConfig
 import io.novafoundation.nova.feature_push_notifications.data.NovaFirebaseMessagingService
 import io.novafoundation.nova.feature_push_notifications.data.data.settings.PushSettings
 import io.novafoundation.nova.feature_push_notifications.data.data.settings.PushSettingsProvider
@@ -37,7 +41,7 @@ class RealPushNotificationsService(
 ) : PushNotificationsService {
 
     init {
-        NovaFirebaseMessagingService.logToken()
+        logToken()
     }
 
     override fun onTokenUpdated(token: String) {
@@ -60,7 +64,7 @@ class RealPushNotificationsService(
     }
 
     override suspend fun onSettingsUpdated(settings: PushSettings): Result<Unit> {
-        if (!googleApiAvailabilityProvider.isAvailable()) return Result.success(Unit)
+        if (!googleApiAvailabilityProvider.isAvailable()) throw IllegalStateException("Google API is not available")
 
         return runCatching {
             if (!isPushNotificationsEnabled()) throw IllegalStateException("Push notifications are not enabled")
@@ -108,5 +112,19 @@ class RealPushNotificationsService(
         }
 
         setNeedToSyncSettings(!succesfullSync)
+    }
+
+    fun logToken() {
+        if (!BuildConfig.DEBUG) return
+
+        FirebaseMessaging.getInstance().token.addOnCompleteListener(
+            OnCompleteListener { task ->
+                if (!task.isSuccessful) {
+                    return@OnCompleteListener
+                }
+
+                Log.d("NOVA_PUSH_TOKEN", "FCM token: ${task.result}")
+            }
+        )
     }
 }
