+++ feature-push-notifications/src/main/java/io/novafoundation/nova/feature_push_notifications/data/data/subscription/RealPushSubscriptionService.kt
@@ -14,6 +14,8 @@ import io.novafoundation.nova.runtime.multiNetwork.chain.model.Chain
 import io.novafoundation.nova.runtime.multiNetwork.chain.model.ChainId
 import io.novafoundation.nova.runtime.multiNetwork.chainsById
 import java.util.*
+import kotlinx.coroutines.sync.Mutex
+import kotlinx.coroutines.sync.withLock
 import kotlinx.coroutines.tasks.await
 
 private const val COLLECTION_NAME = "users"
@@ -25,6 +27,8 @@ class RealPushSubscriptionService(
     private val googleApiAvailabilityProvider: GoogleApiAvailabilityProvider
 ) : PushSubscriptionService {
 
+    private val generateIdMutex = Mutex()
+
     override suspend fun handleSubscription(token: String, pushSettings: PushSettings) {
         if (!googleApiAvailabilityProvider.isAvailable()) return
 
@@ -35,15 +39,17 @@ class RealPushSubscriptionService(
         sendWaletSettingsToFirestore(token, pushSettings)
     }
 
-    private fun getFirestoreUUID(): String {
-        var uuid = prefs.getString(PREFS_FIRESTORE_UUID)
+    private suspend fun getFirestoreUUID(): String {
+        return generateIdMutex.withLock {
+            var uuid = prefs.getString(PREFS_FIRESTORE_UUID)
 
-        if (uuid == null) {
-            uuid = UUID.randomUUID().toString()
-            prefs.putString(PREFS_FIRESTORE_UUID, uuid)
+            if (uuid == null) {
+                uuid = UUID.randomUUID().toString()
+                prefs.putString(PREFS_FIRESTORE_UUID, uuid)
+            }
+
+            uuid
         }
-
-        return uuid
     }
 
     private suspend fun sendWaletSettingsToFirestore(token: String, pushSettings: PushSettings) {
