+++ feature-push-notifications/src/main/java/io/novafoundation/nova/feature_push_notifications/data/di/PushNotificationsFeatureApi.kt
@@ -1,11 +1,12 @@
 package io.novafoundation.nova.feature_push_notifications.data.di
 
+import io.novafoundation.nova.feature_push_notifications.data.NovaFirebaseMessagingService
 import io.novafoundation.nova.feature_push_notifications.data.data.PushNotificationsService
 import io.novafoundation.nova.feature_push_notifications.data.domain.interactor.PushNotificationsInteractor
 
 interface PushNotificationsFeatureApi {
 
-    fun pushNotificationService(): PushNotificationsService
+    fun inject(service: NovaFirebaseMessagingService)
 
     fun pushNotificationInteractor(): PushNotificationsInteractor
 }
