--- feature-push-notifications/src/test/java/io/novafoundation/nova/feature_push_notifications/ExampleUnitTest.kt
@@ -1,5 +1,0 @@
-package io.novafoundation.nova.feature_push_notifications
-
-class ExampleUnitTest {
-
-}
